requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.7.1
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
billiard==3.3.0.23
blinker==1.4
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.1
django-celery==3.3.1
django-celery-beat==2.5.0
django-crontab==0.7.1
django-timezone-field==5.0
djangorestframework==3.14.0
entrypoints==0.3
greenlet==0.4.15
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
install==1.3.5
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
msgpack==0.6.2
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
psycopg==3.1.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing_extensions==4.6.2
tzdata==2023.3
ubuntu-advantage-tools==27.8
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
